+++ src/Faker/DrivingLicense.hs
@@ -1,0 +1,109 @@
+{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE OverloadedStrings #-}
+
+module Faker.DrivingLicense where
+
+import Data.Text (Text)
+import Faker (Fake(..))
+import Faker.Provider.DrivingLicense
+import Faker.TH
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","alabama"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","alaska"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","arizona"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","arkansas"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","california"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","colorado"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","connecticut"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","delaware"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","district_of_columbia"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","florida"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","georgia"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","hawaii"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","idaho"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","illinois"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","indiana"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","iowa"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","kansas"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","kentucky"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","louisiana"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","maine"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","maryland"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","massachusetts"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","michigan"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","minnesota"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","mississippi"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","missouri"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","montana"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","nebraska"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","nevada"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","new_hampshire"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","new_jersey"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","new_mexico"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","new_york"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","north_carolina"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","ohio"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","oklahoma"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","oregon"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","pennsylvania"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","rhode_island"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","south_carolina"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","south_dakota"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","tennessee"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","texas"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","utah"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","vermont"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","virginia"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","washington"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","west_virginia"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","wisconsin"])
+
+$(generateFakeFieldUnresolveds "drivingLicense" ["usa","wyoming"])